added metadata to every page
